CRUD for pills and schema change

+++ src/main/java/pills/models/CategoryModel.java
@@ -1,16 +1,16 @@
 package pills.models;
 
 public class CategoryModel {
-	private long categoryId;
+	private Integer categoryId;
 	private String categoryName;
 	
-	public long getCategoryId(){
+	public Integer getCategoryId(){
 		return categoryId;
 	}
 	public String getCategoryName(){
 		return categoryName;
 	}
-	public void setCategoryId(long categoryId){
+	public void setCategoryId(Integer categoryId){
 		this.categoryId = categoryId;
 	}
 	public void setCategoryName(String categoryName){
