+++ src/main/java/pills/service/CategoryDao.java
@@ -2,12 +2,12 @@ package pills.service;
 
 import java.util.List;
 
-import pills.models.Category;
+import pills.entity.Category;
 
 public interface CategoryDao{
 	public void save(Category category);
 	public void delete(Category category);
 	public List<Category> getAll();
-	public Category getById(long id);
+	public Category getById(Integer id);
 	public void update(Category category);
 }
