+++ src/main/java/pills/service/CategoryService.java
@@ -1,6 +1,6 @@
 package pills.service;
 
-import pills.models.Category;
+import pills.entity.Category;
 import pills.models.CategoryModel;
 
 import java.util.List;
@@ -8,7 +8,7 @@ import java.util.List;
 public interface CategoryService{
 	public CategoryModel addCategory(String category);
 	public void deleteCategory(Category category);
-	public CategoryModel viewCategory(Long id);
+	public CategoryModel viewCategory(Integer id);
 	public List<CategoryModel> viewAll();
 	public void updateCategory(CategoryModel category);
 }
